Set the default color scheme to auto to work with monkay theme
The Packages/BetterFindBuffer/FindResults.hidden-tmTheme will be
the color scheme for find view, but the highlighghwords package
requires all colors it uses to be defined on the
Packages/BetterFindBuffer/FindResults.hidden-tmTheme.
Therefore, the highlighghwords will not work with the
Packages/BetterFindBuffer/FindResults.hidden-tmTheme because it requires
all colors on the highlighghwords to be defined on the
Packages/BetterFindBuffer/FindResults.hidden-tmTheme.